requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.6
django-ckeditor==6.1.0
django-environ==0.7.0
django-jalali==4.3.0
django-js-asset==1.2.2
django-render-partial==0.4
idna==3.2
idpay==0.0.1
jdatetime==3.6.2
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.6

djangorestframework~=3.12.4